requirements.txt
fastapi
uvicorn
pydantic
openai
google-analytics-data
pandas
python-dotenv
